mvL pseudovalidation
pseudovalidate.lassosum.pipeline.R - Modified to be used in a multivariate situation. In order for the changes to work, the following functions needed to be modified.

sd.bfile.R - Modified to be used with our package's lassosum function. As the univariate lassosum function is much faster, we use it if the package is installed.

pseudovalidation.R - Modified to perform destandardization sooner (thus using standardized betas for the computation of bXXb)