From pseudovalide to sumstatsvalidate
sumstatsvalidate requires external correlations to perform validation. A new section in it's code makes sure that SNPs and alleles match between the ls.pipeline object, the test bfile and the new external correlations.

pseudovalidate.lassosum.pipeline and pseudovalidationR are removed. pseudovalidate and pseudovalidation are kept as they will be used in future implementations.