mvL validation
validation.lassosum.pipeline.R - Modified to be used in a multivariate situation.  Phenotype data needs to be provided as a data.frame only.
In order for the changes to work, the following function needed to be modified.

par.pheno.covar.R - Modified to be able to deal with data.frame containing multiple phenotypes.